@@ -2,27 +2,24 @@
   "name": "svelte-app",
   "version": "1.1.0",
   "private": true,
+  "type": "module",
   "scripts": {
     "dev": "vite",
     "build": "vite build",
     "preview": "vite preview"
   },
   "devDependencies": {
-    "@sveltejs/vite-plugin-svelte": "^1.1.0",
-    "prettier": "^2.4.1",
-    "prettier-plugin-svelte": "^2.4.0",
-    "sass": "^1.55.0",
-    "svelte": "^3.52.0",
-    "vite": "^3.2.0"
+    "@sveltejs/vite-plugin-svelte": "^2.0.2",
+    "prettier": "^2.8.3",
+    "prettier-plugin-svelte": "^2.9.0",
+    "sass": "^1.57.1",
+    "svelte": "^3.55.1",
+    "vite": "^4.0.4"
   },
   "dependencies": {
     "@fontsource/ibm-plex-sans": "^4.5.13",
-    "@ibm/plex": "^6.1.1",
-    "@popperjs/core": "^2.11.5",
+    "@popperjs/core": "^2.11.6",
     "animate.css": "^4.1.1",
-    "bootstrap": "^5.1.3",
-    "gsap": "^3.10.4",
-    "sirv-cli": "^2.0.2",
-    "svelte-preprocess": "^4.10.7"
+    "bootstrap": "^5.2.3"
   }
 }
